+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Support for drawing on touch with configurable colors
+* Support for drawing on touch with configurable color and stroke width
 * Suport for image + drawings export with customizable formats
 
 ### Changed